treeNN/requirements.txt
matplotlib==3.5.2
numpy==1.22.4
PyYAML==6.0
torch==1.11.0
torchvision==0.12.0
